Do not print warning if deleting symlink during startup
Bacause we are creating symlinks from the fail2ban folders to the
user defined ones on every startup, we would get warned about that
we are overwriting every file on the second startup. This is because
we are still using the same container, which was modified by the
`entrypoint.sh` scrip the previous start, and thus the symlinks are
already present. Unless you added the `--rm` flag to the Docker run
command, or used `docker-compose down` instead of `stop`, every time
you interacted with this container you would run into this problem.

This change checks if the file is already present, and deletes it,
but will only print an error message if it is NOT a symlink, i.e.
there was a fail2ban default file which had the same name.