Check for valid filter references at startup
Previously we only checked if the log file referenced in a jail.conf
file was present, and attached a `config_watcher` to it if so was
not the case. However, if the filter.conf file, that is referenced in
the same jail.conf file, is not present the fail2ban program would
crash when the jail was started after the log file started to exist.

The added code now checks first if the filter.conf is present in all
jail.conf files, and will exit with code "1" if so is not the case.
This way an admin will be able to identify this problem immediately
instead of when the log file comes online.