Rewrite the `config_watcher` functionality
Previously the code allowed a rename of the jail config file before
the "restart" lock was taken. This meant that it was possible for a
watcher to be able to rename its config file and have this change be
included in a restart which was requested by another watcher. When
the "correct" watcher then obtained the lock, its config file had
already been registered by fail2ban and thus the reload would do
nothing.

Now we made so that the rename and restart both happen while the
watcher is in possession of the lock, so we guarantee that there
is only one jail that has been re-activated every time we restart.
This will make the output from fail2ban a little bit more focused as
well, to only show the changes for this specific jail being activated.