Fix default sender domain used by "actions"
Having `root@fail2ban` as sender will not work since that domain is
not known, and the "sendmail" program will not send anything from
a address which doesn't have a valid DNS record. The reason for this
is that because you would not be able to respond to this address,
since it does not actually exist.

The solution is to just change so it comes from the `hostname`
instead.

https://stackoverflow.com/a/17742092